Rename of SettableTransferXcert to PausableXcert.
